separate routes and controllers
