+++ backend/controllers/userController.js
@@ -1,0 +1,14 @@
+const User = require("../model/user");
+
+exports.getUserByID =async (req, res) => {
+    try {
+      const user = await User.findById(req.params.id);
+      if (!user) {
+        return res.status(404).json({ message: 'User not found' });
+      }
+      res.json(user);
+    } catch (error) {
+      console.error("Error fetching user:", error);
+      res.status(500).json({ message: 'Server error' });
+    }
+  }
